requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
requests
pydantic
